+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Bug fixes:
 * `length<-` would remove attributes from the target even if it was a shared value
 * `length(x) <- N` should not strip attributes if `length(x) == N`, which is not in line with GNU-R documentation,
 but relied upon in the `methods` package #55
+* `as.Date` with invalid date string #56
 
 # 1.0 RC 13